refactor : 데이터 불러오는fetch함수를 최상단 App이 아닌 각각 필요한 컴포넌트로 옮김
