+++ src/component/Nav.js
@@ -1,9 +1,21 @@
 import logoImg from "../assets/img/logo.svg";
 import { LoginButton } from "./Button";
 import style from "../css/Nav.module.css";
-// import "../css/color.css";
+import { useEffect, useState } from "react";
+import { isSignIn } from "../api";
+import "../css/color.css";
 
-function NavProfile({ userInfo }) {
+function NavProfile() {
+  const [userInfo, setUserInfo] = useState("");
+
+  const loadUser = async () => {
+    const userInfo = await isSignIn();
+    setUserInfo(userInfo);
+  };
+
+  useEffect(() => {
+    loadUser();
+  }, []);
   if (userInfo) {
     return (
       <div className={style.profile}>
@@ -19,13 +31,13 @@ function NavProfile({ userInfo }) {
   return <LoginButton />;
 }
 
-function Nav({ userInfo }) {
+function Nav() {
   return (
     <div className={style.root}>
       <a href="/">
         <img className={style.logo} src={logoImg} alt="로고" />
       </a>
-      <NavProfile userInfo={userInfo} />
+      <NavProfile />
     </div>
   );
 }
